updated login layout and updated login/register image stack in smaller devices
